requirements.txt
absl-py==0.7.0
antlr4-python3-runtime==4.8
asttokens==2.4.1
backcall==0.2.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==1.6.0
cmake==3.29.2
configparser==6.0.0
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
docker-pycreds==0.4.0
einops==0.7.0
executing==2.0.1
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.12.2
ftfy==6.1.3
future==0.18.3
gdown==3.13.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==1.0.0
grpcio==1.60.0
gym==0.21.0
gym-minigrid==1.0.3
gym-notices==0.0.8
huggingface-hub==0.20.1
hydra-core==1.0.5
idna==3.6
imageio==2.9.0
imageio-ffmpeg==0.4.5
importlib-metadata==7.0.1
importlib-resources==6.1.1
install==1.3.5
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
kornia==0.4.1
lark==1.1.8
lit==18.1.3
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
meshcat==0.0.18
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.0.6
opencv-python==4.1.2.30
opencv-python-headless==4.2.0.34
packaging==23.2
parso==0.8.3
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.0
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybullet==3.0.4
pycocotools==2.0.7
pyglet==1.5.0
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.0.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sacremoses==0.1.1
scikit-image==0.17.2
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==1.39.1
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
subprocess32==3.5.4
sympy==1.12
tabulate==0.9.0
tdqm==0.0.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tifffile==2023.7.10
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
transformers==4.28.0
transforms3d==0.3.1
triton==2.0.0
typing_extensions @ file:///croot/typing_extensions_1705599297034/work
u-msgpack-python==2.8.0
urllib3==2.1.0
vision-evaluation==0.2.14
wandb==0.10.15
watchdog==0.10.4
wcwidth==0.2.12
Werkzeug==3.0.1
zipp==3.17.0
